+++ app/models/system_log.rb
@@ -1,3 +1,3 @@
 class SystemLog < ApplicationRecord
-  belongs_to :organization
+  belongs_to :organization, optional: true
 end