+++ db/migrate/20200825211849_add_index_to_course_event.rb
@@ -1,0 +1,7 @@
+class AddIndexToCourseEvent < ActiveRecord::Migration[6.0]
+  def change
+    add_index :course_events, :state
+    add_index :course_events, :event_date
+    change_column_null(:system_logs, :organization_id, true)
+  end
+end