+++ spec/spec_helper.rb
@@ -13,6 +13,7 @@
 # it.
 #
 # See http://rubydoc.info/gems/rspec-core/RSpec/Core/Configuration
+require 'simplecov'
 require 'rails_helper'
 
 RSpec.configure do |config|